+++ rust/main/hyperlane-base/src/db/rocks/hyperlane_db.rs
@@ -659,18 +659,18 @@ impl HyperlaneDb for HyperlaneRocksDB {
         self.retrieve_value_by_key(HIGHEST_SEEN_MESSAGE_NONCE, &bool::default())
     }
 
-    fn store_payload_id_by_message_id(
+    fn store_payload_ids_by_message_id(
         &self,
         message_id: &H256,
-        payload_id: &UniqueIdentifier,
+        payload_ids: Vec<UniqueIdentifier>,
     ) -> DbResult<()> {
-        self.store_value_by_key(PAYLOAD_ID_BY_MESSAGE_ID, message_id, payload_id)
+        self.store_value_by_key(PAYLOAD_ID_BY_MESSAGE_ID, message_id, &payload_ids)
     }
 
-    fn retrieve_payload_id_by_message_id(
+    fn retrieve_payload_ids_by_message_id(
         &self,
         message_id: &H256,
-    ) -> DbResult<Option<UniqueIdentifier>> {
+    ) -> DbResult<Option<Vec<UniqueIdentifier>>> {
         self.retrieve_value_by_key(PAYLOAD_ID_BY_MESSAGE_ID, message_id)
     }
 }