+++ typescript/cosmos-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hyperlane-xyz/cosmos-sdk
 
+## 12.3.0
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@12.3.0
+
 ## 12.2.0
 
 ### Patch Changes